requirements.txt
soundfile
hydra-core==1.3.1
scipy==1.10.1
pandas
plotly
librosa
wandb
tqdm
